pyproject.toml
[project]
name = "bingobukku"
version = "0.0.1"
description = ""
authors = [
    {name = "Rodrigo Fernandez",email = "rdr.rkh+github2@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-fasthtml (>=0.12.12,<0.13.0)",
    "sqlmodel (>=0.0.24,<0.0.25)",
    "toml (>=0.10.2,<0.11.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
